[build] fix `PublishPipelineArtifact@1` `condition` (#1149)
I found that `condition` was indented one too far!

    Upload logs
    Skipping step due to condition evaluation.

And so, I wasn't able to retrieve build logs for failed builds!

I could either get out a ruler & unindent the `condition` by two
spaces, or just move it up to be aligned with the block above.

I did a quick review of the other `condition` statements and didn't
find any others to fix.